# docs/backend/openapi.yaml
...
openapi: 3.1.0
info:
  title: AAP 云算接入平台 · 服务端 API
  version: 1.0.0
  description: "供应商端 + 管理端接口。模型定义见 ./json-schema/**；接口清单见 02-API接口模型清单.md。"
servers:
  - url: "http://localhost:8083"
    description: 本地
...
paths:
...
  /quotes:
    get:
      operationId: QT-01
      tags:
        - Quote
      summary: 报价
      description: "接口 ID `QT-01`（依据：真源）"
      x-aap-id: QT-01
      x-aap-source: 真源
      security:
        - bearerAuth: []
      x-aap-roles:
        - PROVIDER
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PageMeta"
                          - type: object
                            properties:
                              list:
                                type: array
                                items:
                                  $ref: "#/components/schemas/QuoteRow"
    post:
      operationId: QT-02
      tags:
        - Quote
      summary: 报价
      description: "接口 ID `QT-02`（依据：真源）"
      x-aap-id: QT-02
      x-aap-source: 真源
      security:
        - bearerAuth: []
      x-aap-roles:
        - PROVIDER
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestQuoteCreate"
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/QuoteDetail"
        4XX:
          description: 业务失败：E-1602,E-1001
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /quotes/{quoteId}:
    get:
      operationId: QT-03
      tags:
        - Quote
      summary: 报价
      description: "接口 ID `QT-03`（依据：真源）"
      x-aap-id: QT-03
      x-aap-source: 真源
      security:
        - bearerAuth: []
      x-aap-roles:
        - PROVIDER
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/QuoteDetail"
        4XX:
          description: 业务失败：E-1401
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: QT-04
      tags:
        - Quote
      summary: 报价
      description: "接口 ID `QT-04`（依据：真源）"
      x-aap-id: QT-04
      x-aap-source: 真源
      security:
        - bearerAuth: []
      x-aap-roles:
        - PROVIDER
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/QuoteDetail"
        4XX:
          description: 业务失败：E-1601
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /quotes/{quoteId}/items:
    post:
      operationId: QT-05
      tags:
        - Quote
      summary: 报价
      description: "接口 ID `QT-05`（依据：真源）"
      x-aap-id: QT-05
      x-aap-source: 真源
      security:
        - bearerAuth: []
      x-aap-roles:
        - PROVIDER
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestQuoteItemsCreate"
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/QuoteDetail"
        4XX:
          description: 业务失败：E-1001,E-1401
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: QT-06
      tags:
        - Quote
      summary: 报价
      description: "接口 ID `QT-06`（依据：真源）"
      x-aap-id: QT-06
      x-aap-source: 真源
      security:
        - bearerAuth: []
      x-aap-roles:
        - PROVIDER
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: quoteId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PageMeta"
                          - type: object
                            properties:
                              list:
                                type: array
                                items:
                                  $ref: "#/components/schemas/QuoteItem"
  /quotes/items/{itemId}:
    get:
      operationId: QT-07
      tags:
        - Quote
      summary: 报价
      description: "接口 ID `QT-07`（依据：真源）"
      x-aap-id: QT-07
      x-aap-source: 真源
      security:
        - bearerAuth: []
      x-aap-roles:
        - PROVIDER
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PageMeta"
                          - type: object
                            properties:
                              list:
                                type: array
                                items:
                                  $ref: "#/components/schemas/QuoteItem"
        4XX:
          description: 业务失败：E-1401
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: QT-08
      tags:
        - Quote
      summary: 报价
      description: "接口 ID `QT-08`（依据：真源）"
      x-aap-id: QT-08
      x-aap-source: 真源
      security:
        - bearerAuth: []
      x-aap-roles:
        - PROVIDER
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestQuoteItemSave"
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PageMeta"
                          - type: object
                            properties:
                              list:
                                type: array
                                items:
                                  $ref: "#/components/schemas/QuoteItem"
        4XX:
          description: 业务失败：E-1001,E-1401,E-1402,E-1403
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /quotes/{quoteId}/submit:
    post:
      operationId: QT-09
      tags:
        - Quote
      summary: 报价
      description: "接口 ID `QT-09`（依据：真源）"
      x-aap-id: QT-09
      x-aap-source: 真源
      security:
        - bearerAuth: []
      x-aap-roles:
        - PROVIDER
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestQuoteSubmit"
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/QuoteDetail"
        4XX:
          description: 业务失败：E-1001,E-1401,E-1402,E-1601,E-1602
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /quotes/{quoteId}/withdraw:
    post:
      operationId: QT-10
      tags:
        - Quote
      summary: 报价
      description: "接口 ID `QT-10`（依据：真源）"
      x-aap-id: QT-10
      x-aap-source: 真源
      security:
        - bearerAuth: []
      x-aap-roles:
        - PROVIDER
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/QuoteDetail"
        4XX:
          description: 业务失败：E-1601
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /quotes/{quoteId}/versions:
    get:
      operationId: QT-11
      tags:
        - Quote
      summary: 报价
      description: "接口 ID `QT-11`（依据：推断）"
      x-aap-id: QT-11
      x-aap-source: 推断
      security:
        - bearerAuth: []
      x-aap-roles:
        - PROVIDER
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: quoteId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PageMeta"
                          - type: object
                            properties:
                              list:
                                type: array
                                items:
                                  $ref: "#/components/schemas/QuoteVersion"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PageMeta:
      $ref: ./json-schema/common/page.schema.json
...
    QuoteItem:
      $ref: ./json-schema/models/quote-item.schema.json
    QuoteRow:
      $ref: ./json-schema/models/quote-row.schema.json
    QuoteDetail:
      $ref: ./json-schema/models/quote-detail.schema.json
    QuoteVersion:
      $ref: ./json-schema/models/quote-version.schema.json
...
    RequestQuoteCreate:
      $ref: ./json-schema/requests/quote-create.schema.json
    RequestQuoteItemsCreate:
      $ref: ./json-schema/requests/quote-items-create.schema.json
    RequestQuoteItemSave:
      $ref: ./json-schema/requests/quote-item-save.schema.json
    RequestQuoteSubmit:
      $ref: ./json-schema/requests/quote-submit.schema.json
...
    Envelope:
      $ref: ./json-schema/common/envelope.schema.json
    Error:
      $ref: ./json-schema/common/error.schema.json
...